% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_building_enrollment}
\alias{download_building_enrollment}
\title{Download building-level enrollment data}
\usage{
download_building_enrollment()
}
\value{
Data frame with building enrollment or empty data frame if unavailable
}
\description{
Download building-level enrollment data
}
\keyword{internal}
