% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_current_enrollment}
\alias{download_current_enrollment}
\title{Download current year enrollment by district}
\usage{
download_current_enrollment()
}
\value{
Data frame with current enrollment data
}
\description{
Downloads the current enrollment file which has more detailed breakdowns.
}
\keyword{internal}
