% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_grade_enrollment}
\alias{download_grade_enrollment}
\title{Download enrollment by grade level}
\usage{
download_grade_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with grade-level enrollment
}
\description{
Download enrollment by grade level
}
\keyword{internal}
