% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_historical_enrollment}
\alias{download_historical_enrollment}
\title{Download historical enrollment summary file}
\usage{
download_historical_enrollment()
}
\value{
List with district and building data frames
}
\description{
Downloads the comprehensive historical enrollment file from Idaho SDE.
This file contains enrollment data for all districts and charter schools
across multiple years.
}
\keyword{internal}
