% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{filter_year_data}
\alias{filter_year_data}
\title{Filter historical data to a specific year}
\usage{
filter_year_data(df, end_year)
}
\arguments{
\item{df}{Data frame from historical file}

\item{end_year}{The school year end to filter to}
}
\value{
Data frame filtered to the requested year
}
\description{
The historical enrollment file has years as columns. This function
identifies the correct year column and filters/pivots as needed.
}
\keyword{internal}
