% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years from Idaho SDE data}
\usage{
get_available_years()
}
\value{
Vector of available school year ends
}
\description{
Get available years from Idaho SDE data
}
\keyword{internal}
