% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from Idaho SDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2024 = 2023-24 school year)}
}
\value{
List with district and building data frames
}
\description{
Downloads enrollment data from Idaho State Department of Education.
Uses historical enrollment files which contain multi-year data.
}
\keyword{internal}
