% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_year_label}
\alias{get_year_label}
\title{Get the academic year label from end_year}
\usage{
get_year_label(end_year)
}
\arguments{
\item{end_year}{The end year of the academic year (e.g., 2024 for 2023-24)}
}
\value{
Character string like "2023-24"
}
\description{
Get the academic year label from end_year
}
\keyword{internal}
