% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R, R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years from Idaho SDE data}
\usage{
get_available_years()

get_available_years()
}
\value{
Vector of available school year ends

A list with components:
\describe{
\item{min_year}{Earliest available year (2002)}
\item{max_year}{Most recent available year (2024)}
\item{description}{Human-readable description of the date range}
}
}
\description{
Returns the range of years for which enrollment data can be fetched
from the Idaho State Department of Education.
}
\examples{
get_available_years()
}
\keyword{internal}
